requirements.txt
Flask== 1.1.4
gevent== 20.9.0
gunicorn== 20.0.4
redis== 3.5.3
PyMySQL== 0.10.1
APScheduler==3.6.3
cleo== 0.6.8
pastel== 0.1.1
requests== 2.24.0
clikit== 0.2.4
flask-redis== 0.4.0
Flask-Cors== 3.0.9
flask-orator== 0.2.0
Flask-Sockets== 0.2.1
gevent-websocket== 0.10.1
supervisor== 4.2.1
loguru== 0.5.3
elasticsearch== 7.9.1
python-nmap== 0.6.1
pytz== 2020.4
OTXv2== 1.5.10
jira== 2.0.0
paramiko== 2.7.2
pywinrm== 0.4.1
rpyc== 5.0.0
nest-asyncio== 1.5.4